Update gql logic to use repositories
