+++ src/server/service/graphql/schema/context.rs
@@ -1,0 +1,3 @@
+use crate::server::data::Registry;
+
+impl juniper::Context for Registry {}
