+++ src/server/service/graphql/schema/subscriptions.rs
@@ -1,0 +1,3 @@
+use crate::server::data::Registry;
+
+pub type Subscriptions = juniper::EmptySubscription<Registry>;
