--- src/server/services/graphql/schema/context.rs
@@ -1,3 +1,0 @@
-use crate::database::connection::DatabaseConnection;
-
-impl juniper::Context for DatabaseConnection {}
