--- src/server/services/graphql/schema/subscriptions.rs
@@ -1,3 +1,0 @@
-use crate::database::connection::DatabaseConnection;
-
-pub type Subscriptions = juniper::EmptySubscription<DatabaseConnection>;
